+++ concordium-consensus/src/Concordium/GlobalState/Persistent/TreeState.hs
@@ -1,12 +1,12 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE MultiWayIf #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# LANGUAGE MultiWayIf #-}
 -- FIXME: This is to suppress compiler warnings for derived instances of BlockStateOperations.
 -- This may be fixed in GHC 9.0.1.
 {-# OPTIONS_GHC -Wno-redundant-constraints #-}
@@ -47,11 +47,11 @@ import qualified Data.Sequence as Seq
 import Lens.Micro.Platform
 import Concordium.Utils
 import System.Mem.Weak
-import System.Directory
-import System.IO.Error
-import System.FilePath
 import Concordium.Logger
 import Control.Monad.Except
+import System.FilePath
+import System.Directory
+import System.IO.Error
 import qualified Concordium.TransactionVerification as TVer
 
 -- * Exceptions
@@ -306,9 +306,9 @@ initialSkovPersistentData rp treeStateDir gd genState serState genTT mPending = 
 
 -- * Initialization functions
 
--- |Check the permissions in the required files.
--- Returns 'True' if the database already existed, 'False' if not.
--- Raises an exception if the database is inaccessible, or only partially exists.
+--- |Check the permissions in the required files.  Returns 'True' if the database already exists,
+--- 'False' if it does not exist, or is inaccessible.  If the database exists only partially, then it
+--- is deleted to allow for creating it again through `newGenesis`.
 checkExistingDatabase :: forall m. (MonadLogger m, MonadIO m) =>
     -- |Tree state path
     FilePath ->
@@ -343,9 +343,13 @@ checkExistingDatabase treeStateDir blockStateFile = do
          mapM_ (logEvent TreeState LLTrace) ["Existing database found.", "TreeState filepath: " ++ show blockStateFile, "BlockState filepath: " ++ show treeStateFile]
          return True
      | bsPathEx -> do
-         logExceptionAndThrowTS $ DatabaseInvariantViolation "Block state file exists, but tree state file does not."
+         logEvent GlobalState LLWarning "Block state file exists, but tree state database does not. Deleting the block state file."
+         liftIO $ removeFile blockStateFile
+         return False
      | tsPathEx -> do
-         logExceptionAndThrowTS $ DatabaseInvariantViolation "Tree state file exists, but block state file does not."
+         logEvent GlobalState LLWarning "Tree state database exists, but block state file does not. Deleting the tree state database."
+         liftIO . removeDirectoryRecursive $ treeStateDir
+         return False
      | otherwise ->
          return False
 
@@ -386,23 +390,22 @@ loadSkovPersistentData rp _treeStateDirectory pbsc = do
   liftIO (checkDatabaseVersion _db) >>=
       either (logExceptionAndThrowTS . IncorrectDatabaseVersion) return
 
-  -- Get the genesis block and check that its data matches the supplied genesis data.
-  genStoredBlock <- maybe (logExceptionAndThrowTS GenesisBlockNotInDataBaseError) return =<<
-          liftIO (getFirstBlock _db)
-  _genesisBlockPointer <- liftIO $ makeBlockPointer genStoredBlock
-  _genesisData <- case _bpBlock _genesisBlockPointer of
-    GenesisBlock gd' -> return gd'
-    _ -> logExceptionAndThrowTS (DatabaseInvariantViolation "Block at height 0 is not a genesis block.")
-
   -- Unroll the treestate if the last finalized blockstate is corrupted. If the last finalized
   -- blockstate is not corrupted, the treestate is unchanged.
   unrollTreeStateWhile _db (liftIO . isBlockStateCorrupted) >>= \case
     Left e -> logExceptionAndThrowTS . DatabaseInvariantViolation $
               "The block state database is corrupt. Recovery attempt failed: " <> e
     Right (_lastFinalizationRecord, lfStoredBlock) -> do
-      -- Truncate the blobstore beyond the last finalized blockstate. If no corruption was found
-      -- earlier, the blobstore size does not change.
+      -- Truncate the blobstore beyond the last finalized blockstate.
       liftIO $ truncateBlobStore (bscBlobStore . PBS.pbscBlobStore $ pbsc) (sbState lfStoredBlock)
+      -- Get the genesis block.
+      genStoredBlock <- maybe (logExceptionAndThrowTS GenesisBlockNotInDataBaseError) return =<<
+              liftIO (getFirstBlock _db)
+      _genesisBlockPointer <- liftIO $ makeBlockPointer genStoredBlock
+      _genesisData <- case _bpBlock _genesisBlockPointer of
+        GenesisBlock gd' -> return gd'
+        _ -> logExceptionAndThrowTS (DatabaseInvariantViolation "Block at height 0 is not a genesis block.")
+      -- Get the last finalized block.
       _lastFinalized <- liftIO (makeBlockPointer lfStoredBlock)
       return SkovPersistentData {
             _possiblyPendingTable = HM.empty,