+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## Unreleased changes
 
+- Fix a bug in database recovery where the node would hang when truncating the block state database
+  on Windows.
+
 ## 4.4.3
 
 - Fix a bug in database recovery where corruption was not always correctly detected.
 - Fix typo in environment variable `CONCORDIUM_NODE_PROMETHEUS_LISTEN_ADDRESSS` (remove trailing `S`).
+- The node is now able to recover after crashes which leave only treestate or only blockstate usable.
 
 ## 4.4.2